修修改改

+++ app/Services/ResumeServices.php
@@ -59,7 +59,8 @@ class ResumeServices
         }
     }
 
-    public function newAbilityById($re){
+    public function newAbilityById($re)
+    {
         $token = JWTAuth::getToken();
         $user = JWTAuth::toUser($token);
         $id = $user->id;
@@ -79,7 +80,7 @@ class ResumeServices
     //新增履歷結束
 
     //修改履歷開始
-    public function editBasicDataById_ser($request,$file,$l_file)
+    public function editBasicDataById_ser($request, $file, $l_file)
     {
 
         $re = $request->all();
@@ -88,48 +89,48 @@ class ResumeServices
         $id = $user->id;
         $stuBas = stuBasicEloquent::where('sid', $id)->first();
 
-        if($file){
-        $extension = $file->getClientOriginalExtension();
-        $file_name = strval(time()).str_random(5).'_pro.'.$extension;
+        if ($file) {
+            $extension = $file->getClientOriginalExtension();
+            $file_name = strval(time()) . str_random(5) . '_pro.' . $extension;
 
-        if ($request->hasFile('profilePic')) {
-            if($stuBas->profilePic!=null){
+            if ($request->hasFile('profilePic')) {
+                if ($stuBas->profilePic != null) {
 
-                $file_path = 'public/user-upload/'.$stuBas->profilePic;
-                $file=Storage::exists('public/user-upload/'.$stuBas->profilePic);
+                    $file_path = 'public/user-upload/' . $stuBas->profilePic;
+                    $file = Storage::exists('public/user-upload/' . $stuBas->profilePic);
 
-                Log::error(Storage::exists('public/user-upload/'.$stuBas->profilePic));
+                    Log::error(Storage::exists('public/user-upload/' . $stuBas->profilePic));
 
-                if($file){
-                    Storage::delete($file_path);
-                }else{
-                    return 'failed';
+                    if ($file) {
+                        Storage::delete($file_path);
+                    } else {
+                        return 'failed';
+                    }
                 }
+                $path = $request->file('profilePic')->storeAs(
+                    'public/user-upload/', $file_name
+                );
+                //<img src='storage/user-upload/1501257619SWUxK.png' >
+                $stuBas->profilePic = $file_name;
+            } else {
+                return "頭貼上傳失敗";
             }
-            $path = $request->file('profilePic')->storeAs(
-                'public/user-upload/', $file_name
-            );
-            //<img src='storage/user-upload/1501257619SWUxK.png' >
-            $stuBas->profilePic=$file_name;
-        } else {
-            return "頭貼上傳失敗";
         }
-    }
-        if($l_file){
+        if ($l_file) {
             $extension = $l_file->getClientOriginalExtension();
-            $file_name = strval(time()).str_random(5).'_lic.'.$extension;
+            $file_name = strval(time()) . str_random(5) . '_lic.' . $extension;
 
             if ($request->hasFile('licenceFile')) {
-                if($stuBas->profilePic!=null){
+                if ($stuBas->profilePic != null) {
 
-                    $file_path = 'public/user-upload/licences/'.$stuBas->licenceFile;
-                    $l_file=Storage::exists('public/user-upload/licences/'.$stuBas->licenceFile);
+                    $file_path = 'public/user-upload/licences/' . $stuBas->licenceFile;
+                    $l_file = Storage::exists('public/user-upload/licences/' . $stuBas->licenceFile);
 
-                    Log::error(Storage::exists('public/user-upload/licences/'.$stuBas->licenceFile));
+                    Log::error(Storage::exists('public/user-upload/licences/' . $stuBas->licenceFile));
 
-                    if($l_file){
+                    if ($l_file) {
                         Storage::delete($file_path);
-                    }else{
+                    } else {
                         return 'failed';
                     }
                 }
@@ -137,11 +138,18 @@ class ResumeServices
                     'public/user-upload/licences/', $file_name
                 );
                 //<img src='storage/user-upload/licences/1501257619SWUxK.png' >
-                $stuBas->licenceFile=$file_name;
+                $stuBas->licenceFile = $file_name;
             } else {
                 return "證照檔案上傳失敗";
             }
         }
+        if (isset($re['Oname'])) {
+            $stuBas->Oname = $re['Oname'];
+            $stuBas->OS = $re['OS'];
+            $stuBas->OR = $re['OR'];
+            $stuBas->OW = $re['OW'];
+        } elseif (isset($re['OS']) || isset($re['OR']) || isset($re['OW']))
+            return '請填寫其他語言名稱';
         $stuBas->chiName = $re['chiName'];
         $stuBas->engName = $re['engName'];
         $stuBas->bornedPlace = $re['bornedPlace'];
@@ -153,18 +161,18 @@ class ResumeServices
         $stuBas->ES = $re['ES'];
         $stuBas->ER = $re['ER'];
         $stuBas->EW = $re['EW'];
-        $stuBas->TOEIC = $re['TOEIC'];
-        $stuBas->TOEFL = $re['TOEFL'];
+        if (isset($re['TOEIC']))
+            $stuBas->TOEIC = $re['TOEIC'];
+        if (isset($re['TOEFL']))
+            $stuBas->TOEFL = $re['TOEFL'];
         $stuBas->autobiography = $re['autobiography'];
-        $stuBas->Oname = $re['Oname'];
-        $stuBas->OS = $re['OS'];
-        $stuBas->OR = $re['OR'];
-        $stuBas->OW = $re['OW'];
         $stuBas->eduSystem = $re['eduSystem'];
         $stuBas->graduateYear = $re['graduateYear'];
         $stuBas->graduatedSchool = $re['graduatedSchool'];
-        $stuBas->department = $re['department'];
-        $stuBas->section = $re['section'];
+        if (isset($re['department']))
+            $stuBas->department = $re['department'];
+        if (isset($re['section']))
+            $stuBas->section = $re['section'];
         $stuBas->save();
         if (stuBasicEloquent::count() != 0) { //rowcount
             return '修改基本資料成功';
@@ -174,39 +182,41 @@ class ResumeServices
     }
 
 
-
-
     public function deleteJobExperienceById_ser($re)
     {
         $stuJExp = stuJExpEloquent::where('jid', $re['jid'])->first();
-        $stuJExp->delete();
-        if (stuJExpEloquent::count() != 0) {
+        if ($stuJExp) {
+            $stuJExp->delete();
             return '刪除工作資料成功';
         } else {
-            return '刪除工作資料失敗';
+            return '查無此資料';
         }
+
     }
 
 
     public function deleteWorksDataById_ser($re)
     {
         $stuWor = stuWorksEloquent::where('wid', $re['wid'])->first();
-        $stuWor->delete();
-        if (stuWorksEloquent::count() != 0) {
+        if ($stuWor) {
+            $stuWor->delete();
             return '刪除作品資料成功';
         } else {
-            return '刪除作品資料失敗';
+            return '查無此資料';
         }
+
     }
 
-    public function deleteAbilityById_ser($re){
+    public function deleteAbilityById_ser($re)
+    {
         $stuA = stuAbilityEloquent::where('abiid', $re['abiid'])->first();
-        $stuA->delete();
-        if (stuAbilityEloquent::count() != 0) {
+        if ($stuA) {
+            $stuA->delete();
             return '刪除能力資料成功';
         } else {
-            return '刪除能力資料失敗';
+            return '查無此資料';
         }
+
     }
     //修改履歷結束
 }
